requirements.txt
altair==4.0.0
langchain==0.0.234
openai==0.27.8
streamlit==1.12.0
wikipedia==1.4.0
python-dotenv
